remvoe cosnensus/no_running_consensus_instance test
In the past it tested cases in which a consensus message is received but no running qbft instance is present.
After a few changes in the code (over the past few months) in such cases it will either be a future message or carrying a pre-consensus in which case a qbft instance will be started.

Essentially it doesn't test anything anymore